Fix "$hitMeButton is not defined"
This was failing because the call was outside the $(document).ready()
closure.

Because app.js is loaded in the HTML document header, the script was
being run prior to the hit me button actually existing. By moving these
lines up into the document ready closure, we ensure that the lines
aren't run until the whole page is fully loaded.